+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - `finage` to get Financial data from finage.co.uk
   - `coincodex` to get crypto prices from CoinCodex
   - `coinranking` to get crypto prices from Coinranking
+  - `crypto-volatility-index` to calculate the CVI (Crypto volatility index)
   - `btc.com` to get on-chain balances from BTC.com
   - `sochain` to get on-chain balances from SoChain
   - `dns-query` to query DNS over HTTPS
@@ -32,6 +33,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add a package re-usable adapter factory implementations to - `@chainlink/ea-factories`
 
 - Added support for using multiple API Keys per-adapter
+
+### Fixed
+
+- CMC adapter price query fixed for currencies where multiple currencies have the same symbol
 
 ### Changed